+++ source/problem/SWE/problem_postprocessor/swe_post_write_vtu.hpp
@@ -32,22 +32,22 @@ void write_VTU_data(MeshType& mesh, std::ofstream& raw_data_file) {
 
     raw_data_file << "\t\t\t\t<DataArray type=\"Float64\" Name=\"ze_point\" format=\"ascii\">\n\t\t\t\t\t";
     for (auto it = q_point_data.begin(); it != q_point_data.end(); ++it)
-        raw_data_file << (*it)[SWE::Variables::ze] << ' ';
+        raw_data_file << (float)(*it)[SWE::Variables::ze] << ' ';
     raw_data_file << "\n\t\t\t\t</DataArray>\n";
 
     raw_data_file << "\t\t\t\t<DataArray type=\"Float64\" Name=\"qx_point\" format=\"ascii\">\n\t\t\t\t\t";
     for (auto it = q_point_data.begin(); it != q_point_data.end(); ++it)
-        raw_data_file << (*it)[SWE::Variables::qx] << ' ';
+        raw_data_file << (float)(*it)[SWE::Variables::qx] << ' ';
     raw_data_file << "\n\t\t\t\t</DataArray>\n";
 
     raw_data_file << "\t\t\t\t<DataArray type=\"Float64\" Name=\"qy_point\" format=\"ascii\">\n\t\t\t\t\t";
     for (auto it = q_point_data.begin(); it != q_point_data.end(); ++it)
-        raw_data_file << (*it)[SWE::Variables::qy] << ' ';
+        raw_data_file << (float)(*it)[SWE::Variables::qy] << ' ';
     raw_data_file << "\n\t\t\t\t</DataArray>\n";
 
     raw_data_file << "\t\t\t\t<DataArray type=\"Float64\" Name=\"bath_point\" format=\"ascii\">\n\t\t\t\t\t";
     for (auto it = aux_point_data.begin(); it != aux_point_data.end(); ++it)
-        raw_data_file << (*it)[SWE::Auxiliaries::bath] << ' ';
+        raw_data_file << (float)(*it)[SWE::Auxiliaries::bath] << ' ';
     raw_data_file << "\n\t\t\t\t</DataArray>\n";
 
     raw_data_file << "\t\t\t</PointData>\n";
@@ -56,22 +56,22 @@ void write_VTU_data(MeshType& mesh, std::ofstream& raw_data_file) {
 
     raw_data_file << "\t\t\t\t<DataArray type=\"Float64\" Name=\"ze_cell\" format=\"ascii\">\n\t\t\t\t\t";
     for (auto it = q_cell_data.begin(); it != q_cell_data.end(); ++it)
-        raw_data_file << (*it)[SWE::Variables::ze] << ' ';
+        raw_data_file << (float)(*it)[SWE::Variables::ze] << ' ';
     raw_data_file << "\n\t\t\t\t</DataArray>\n";
 
     raw_data_file << "\t\t\t\t<DataArray type=\"Float64\" Name=\"qx_cell\" format=\"ascii\">\n\t\t\t\t\t";
     for (auto it = q_cell_data.begin(); it != q_cell_data.end(); ++it)
-        raw_data_file << (*it)[SWE::Variables::qx] << ' ';
+        raw_data_file << (float)(*it)[SWE::Variables::qx] << ' ';
     raw_data_file << "\n\t\t\t\t</DataArray>\n";
 
     raw_data_file << "\t\t\t\t<DataArray type=\"Float64\" Name=\"qy_cell\" format=\"ascii\">\n\t\t\t\t\t";
     for (auto it = q_cell_data.begin(); it != q_cell_data.end(); ++it)
-        raw_data_file << (*it)[SWE::Variables::qy] << ' ';
+        raw_data_file << (float)(*it)[SWE::Variables::qy] << ' ';
     raw_data_file << "\n\t\t\t\t</DataArray>\n";
 
     raw_data_file << "\t\t\t\t<DataArray type=\"Float64\" Name=\"bath_cell\" format=\"ascii\">\n\t\t\t\t\t";
     for (auto it = aux_cell_data.begin(); it != aux_cell_data.end(); ++it)
-        raw_data_file << (*it)[SWE::Auxiliaries::bath] << ' ';
+        raw_data_file << (float)(*it)[SWE::Auxiliaries::bath] << ' ';
     raw_data_file << "\n\t\t\t\t</DataArray>\n";
 
     raw_data_file << "\t\t\t\t<DataArray type=\"UInt32\" Name=\"ID\" format=\"ascii\">\n\t\t\t\t\t";
